Moved collectionsDropdown to it's own class
